API: Problems & Dynamics

Continuous problems

Main.SCPLib.get_trajectory_augmented — Method

Propagate augmented dynamics with continuous control This function also constructs state-transition matrices and evaluates dynamics residuals.

Main.SCPLib.get_trajectory_augmented_forwardbackward — Method

Propagate augmented dynamics with continuous control This function also constructs state-transition matrices and evaluates dynamics residual at the match point.

Main.SCPLib.init_continuous_dynamics_xaug — Method

Initialize augmented state for continuous dynamics

Main.SCPLib.ContinuousProblem — Method

Construct a continuous optimal control problem

If set_dynamics_cache! is provided, it will be used to set STM's to prob.lincache. The signature of set_dynamics_cache! should be:

set_dynamics_cache!(prob::OptimalControlProblem, x_ref::Union{Matrix,Adjoint}, u_ref::Union{Matrix,Adjoint}, y_ref::Union{Matrix,Nothing})

and the function should return the dynamics residuals. See set_dynamics_cache! for more details.

Arguments:

optimizer: optimizer for the JuMP model

eom!::Function: continuous dynamics function

params: parameters for the problem

objective::Function: objective function

times: time points

x_ref: reference state

u_ref: reference control

eom_aug!::Function: augmented dynamics function

ng::Int: number of non-convex equality constraints

g_noncvx::Union{Function,Nothing}: non-convex equality constraints with signature (lincache, x, u) -> g

∇g_noncvx::Union{Function,Nothing}: gradient of non-convex equality constraints with signature (lincache, x, u) -> ∇g

nh::Int: number of non-convex inequality constraints

h_noncvx::Union{Function,Nothing}: non-convex inequality constraints with signature (lincache, x, u) -> h

∇h_noncvx::Union{Function,Nothing}: gradient of non-convex inequality constraints with signature (lincache, x, u) -> ∇h

ode_ensemble_method: ensemble method for the ODE solver

ode_method: method for the ODE solver

ode_reltol: relative tolerance for the ODE solver

ode_abstol: absolute tolerance for the ODE solver

fun_get_trajectory::Union{Function,Nothing}: user-defined function to get the trajectory

set_dynamics_cache!::Union{Function,Nothing}: user-defined function to set the dynamics cache

u_bias::Union{Matrix,Nothing}: bias on the control

Impulsive problems

Main.SCPLib.get_trajectory_augmented — Method

Propagate augmented dynamics with continuous control This function also constructs state-transition matrices and evaluates dynamics residuals.

Main.SCPLib.init_impulsive_dynamics_xaug — Method

Initialize augmented state for continuous dynamics

Main.SCPLib.ImpulsiveProblem — Method

Construct an impulsive optimal control problem

If set_dynamics_cache! is provided, it will be used to set STM's to prob.lincache. The signature of set_dynamics_cache! should be:

set_dynamics_cache!(prob::OptimalControlProblem, x_ref::Union{Matrix,Adjoint}, u_ref::Union{Matrix,Adjoint}, y_ref::Union{Matrix,Nothing})

and the function should return the dynamics residuals. See set_dynamics_cache! for more details.

Arguments:

optimizer: optimizer for the JuMP model

eom!::Function: impulsive dynamics function

params: parameters for the problem

objective::Function: objective function

times: time points

x_ref: reference state

u_ref: reference control

eom_aug!::Function: augmented dynamics function

dfdu::Function: derivative of the impulsive dynamics function with respect to the control

ng::Int: number of non-convex equality constraints

g_noncvx::Union{Function,Nothing}: non-convex equality constraints with signature (lincache, x, u) -> g

∇g_noncvx::Union{Function,Nothing}: gradient of non-convex equality constraints with signature (lincache, x, u) -> ∇g

nh::Int: number of non-convex inequality constraints

h_noncvx::Union{Function,Nothing}: non-convex inequality constraints with signature (lincache, x, u) -> h

∇h_noncvx::Union{Function,Nothing}: gradient of non-convex inequality constraints with signature (lincache, x, u) -> ∇h

ode_ensemble_method: ensemble method for the ODE solver

ode_method: method for the ODE solver

ode_reltol: relative tolerance for the ODE solver

ode_abstol: absolute tolerance for the ODE solver

fun_get_trajectory::Union{Function,Nothing}: user-defined function to get the trajectory

set_dynamics_cache!::Union{Function,Nothing}: user-defined function to set the dynamics cache

u_bias::Union{Matrix,Nothing}: bias on the control

Equations of motion

Main.SCPLib.get_continuous_augmented_eom — Method

Get augmented in-place dynamics function with signature eom_aug!(dx_aug, x_aug, p, t)

Main.SCPLib.get_continuous_augmented_eom — Method

Get augmented in-place dynamics function with signature eom_aug!(dx_aug, x_aug, p, t)

Main.SCPLib.get_impulsive_augmented_eom — Method

Get augmented in-place dynamics function with signature eom_aug!(dx_aug, x_aug, p, t)

Main.SCPLib.get_impulsive_augmented_eom — Method

Get augmented in-place dynamics function with signature eom_aug!(dx_aug, x_aug, p, t)