API: Problems & Dynamics

Continuous problems

Main.SCPLib.get_trajectory_augmented — Method

Propagate augmented dynamics with continuous control This function also constructs state-transition matrices and evaluates dynamics residuals.

Main.SCPLib.init_continuous_dynamics_xaug — Method

Initialize augmented state for continuous dynamics

Main.SCPLib.ContinuousProblem — Method

Construct a continuous optimal control problem

Impulsive problems

Main.SCPLib.get_trajectory_augmented — Method

Propagate augmented dynamics with continuous control This function also constructs state-transition matrices and evaluates dynamics residuals.

Main.SCPLib.init_impulsive_dynamics_xaug — Method

Initialize augmented state for continuous dynamics

Main.SCPLib.ImpulsiveProblem — Method

Construct an impulsive optimal control problem

If set_dynamics_cache! is provided, it will be used to set STM's to prob.lincache.

The signature of set_dynamics_cache! should be:

set_dynamics_cache!(prob::OptimalControlProblem, x_ref::Union{Matrix,Adjoint}, u_ref::Union{Matrix,Adjoint}, y_ref::Union{Matrix,Nothing})

and the function should return the dynamics residuals. See set_dynamics_cache! for more details.

Equations of motion

Main.SCPLib.get_continuous_augmented_eom — Method

Get augmented in-place dynamics function with signature eom_aug!(dx_aug, x_aug, p, t)

Main.SCPLib.get_continuous_augmented_eom — Method

Get augmented in-place dynamics function with signature eom_aug!(dx_aug, x_aug, p, t)

Main.SCPLib.get_impulsive_augmented_eom — Method

Get augmented in-place dynamics function with signature eom_aug!(dx_aug, x_aug, p, t)

Main.SCPLib.get_impulsive_augmented_eom — Method

Get augmented in-place dynamics function with signature eom_aug!(dx_aug, x_aug, p, t)